Update cli go
